linting/error handling
