+++ receiver/awsefareceiver/generated_package_test.go
@@ -3,8 +3,9 @@
 package awsefareceiver
 
 import (
-	"go.uber.org/goleak"
 	"testing"
+
+	"go.uber.org/goleak"
 )
 
 func TestMain(m *testing.M) {
